Maude Alpha86e
Dear Maude Abusers,

This version is a Maude 2.2 release candidate and reports itself as
such. I have made binaries for the same 5 targets as I did for
alpha86d. If anyone needs a binary for some other target speak now.

This version exists to fix a nasty bug reported by Paco and which
exists in all Maude 2 versions. The problem is that the metalevel
fails to record prec attributes properly, and while they work, there
is a problem if a metalevel declared operator subsort overloads an
object level operator which has the prec attribute:

fmod FOO is
  op f_ : Bool -> Bool [prec 14] .
endfm

red in META-LEVEL : metaReduce(
(
fmod 'BAR is
  protecting 'FOO .
  sorts none .
  none
  op 'f_ : 'Bool -> 'Bool [prec(14)] .
  none
  none
endfm
),
'true.Bool) .

Steven