Maude Alpha87
Dear Maude Abusers,

This version is a pure alpha release - in fact the only reason for
making an alpha release at this time is to have the strategy language
fragment available for the Maude demo at WRLA 2006.

New feature
============
There is very crude support for a fragment of the strategy
language. It is envoked by the command:

srewrite [<nat>] in <module> : <term> using <strategy> .

Here the
  [<nat>]
(which specifies the number of solutions to find) and
  in <module> :
parts are optional. srewrite is a bit like search in that looks for
sequences of rewrites, but here the restriction on a sequence is that
it must satisfy the strategy.

Strategies are formed as follows; the leaf strategies are:

fail	   produce the empty set of successors
idle	   produce the identity rewrite
all	   apply any rule
l	   apply a rule with label l
l[s]	   apply a rule with label l and substitution s

Substitions are a comma seperated list:
v1 <- t1 ,..., vk <- tk

The strategy combinators are:

s *	   do s 0 or more times
s +	   do s 1 or more times
s ; s'	   do s and then s'
s | s'	   do s or s'
t ? s : f  do t, if it produces results to s on them else do f on the
	   original term

Note that at each step a strategy transforms a term into a set of
successors.

Limitations
------------
(a) Much of the proposed Maude strategy language remains to be
implemented; there are no strategy modules.
(b) The same answer may be produced many times if the are many ways to
produce it - there is no way to cache previous results and exclude
duplicates.
(c) There is no way to continue (the continue command does not work
for srewrite).
(d) The is no attempt at efficiency.
(e) The command echoing for srewrite prints raw tokens as there is no
print routine for strategies as yet.

Minor change
=============
"show subsorts ." now displays the main sort in each subsort
declaration in magenta.

Steven