Maude Alpha87b
Dear Maude Abusers,

This release exists to fix 2 bugs:

(1) A bug in the implementation of the ! strategy combinator that can
    cause it to go into a infinite loop:

mod MAP is
   sorts Elem List .
   subsort Elem < List .
   ops a b c : -> Elem .
   op nil : -> List .
   op __ : List List -> List [assoc id: nil] .
   rl [g] : b => c .
endm

srewrite a a using g ! .

    This code is simplified from the supplied by Alberto who reported
    the bug.

(2) A nasty bug in the symbol look up code for associative ops that
    has existed for 7.5 years and is exposed by this example from Joe
    Hendrix:

select META-LEVEL .
red maximalAritySet(['BOOL], '_and_, '`[Bool`] '`[Bool`] '`[Bool`], 'Bool) .

Also this version fixed configure.ac for building on Intel Macs.

Finally I have gone back to gcc 4.0.1 on RedHat 7.3 for the linux
version but now instead of being a static binary, some libraries which
are OS dependent but have a stable API _and_ a stable ABI (no compiler
dependencies) are dynamically linked. This is another attempt to fix
the gethostbyname() problem reported by Matthias Radestock.

Steven