Maude Alpha88
Dear Maude Abusers,

The main reason for a new release at this time is Jose's need for new
feature (1).

Bug fix
========
Two missing ctor bugs found by Peter and reported by Narcisso:
In META-MODULE
   op _{_}  : Qid ParameterDeclList -> Header .
In MODEL-CHECKER
   op {_,_} : State RuleName -> Transition .

New features
=============
(1) The search command can now take a depth bound; for example:

mod SEARCH-TEST is
  pr NAT .
  sort List .
  subsort Nat < List .
  op __ : List List -> List [assoc] .
  var N : Nat .
  var L : List .
  rl N => N + 1 .
endm

search [100, 8] 0 0 0 =>* L .

It is also possible to give a depth bound without giving a bound on
the number of solutions returned:

search [, 8] 0 0 0 =>* L .

Note the as the code for handling double numeric arguments is shared
with frewrite/erewrite, this syntax can now be used to give a
non-unity value of gas without giving a bound on the number of
rewrites. This doesn't seem particularly useful but it is easier to
document this behavior that complicate the already extremely
complicated surface parser to remove it. Note that gas doesn't work
for erewrite at the moment anyway.

(2) The is a new descent function
  op metaNormalize : Module Term ~> ResultPair [special (...)] .
that puts a term in theory normal form without doing any rewrites. For
example

red in META-LEVEL : metaNormalize(upModule('NAT, false),
  's_['s_['0.Zero]]) .

yields {'s_^2['0.Zero],'NzNat} . This feature was requested by Paco.

(3) The implementation of Maude 1.0 style machine integers in Maude
that Jose and I have been playing with for a while is now part of the
distribution. To use it:

load machine-int

fmod TEST is
  inc MACHINE-INT{32-BIT} .
endfm

Wrap-around from negative to positive:

Maude> red -2147483648 - 1 .
result NzMachineNat: 2147483647

Wrap-around from positive to negative:

Maude> red 2147483647 + 1 .
result NzMachineInt: -2147483648

64 bit integers are likewise available and as are k-bit integers for
any k >= 2 that is a power of 2 by defining an appropriate view from
BIT-WIDTH to RENAMED-INT.

Steven