Maude Alpha88c
Dear Maude Abusers,

This version reports itself as Maude 2.3 and is a 2.3 release
candidate to go with the forthcoming Maude book.

Bug fixes
==========
(1) The projection functions for metamodules have been updated to
support parameterized metamodules. Bug reported by Joe Hendrix.

(2) A bug where all but the first sort in a kind was printed twice by
metaPrettyPrint(). Provoked by the following example.

red in META-LEVEL : metaPrettyPrint(['META-LEVEL], ''V:`[Term`]) .

Bug reported by Joe Hendrix.

Changes
========
(1) The =>+ version of the search command now allows the initial state
to be matched if there is a cycle of length >= 1 back to it. This also
affects the deprecated =>1 version since this is just an alternative
notation for =>+ with a depth bound of 1. The '+ option in
metaSearch() is likewise affected. Requested by Jose.

(2) min() and max() operations are added to FLOAT. They work with
infinities as well as FiniteFloats. Requested by Marc Boyer
<marc.boyer@enseeiht.fr>.

(3) Use of non-existent or unpatchable modules at the metalevel now
generates an advisory. Requested by Joe Hendrix.

(4) erewrite now respects the number of rewrites bound and supports
continue so for example with the web client code we can do:

erew [1, 2] start("www.google.com") .

cont 1 .

cont .

As with rewrite and frewrite, the bound only applies to regular top
level rule rewrites; equation and membership applications and
searching in rewrite conditions is unbounded.

Finally as an addendum to the alpha88b release notes, that version
added support for continue to srewrite, if you give a bound on the
number of solutions you asked for and then want more.

Steven