Maude Alpha89b
Dear Maude Abusers,

This release is to fix a critical bug in the unification
implementation where under certain circumstance the nonexistence of
unifiers could lead to a pointer not being initialized and a random
piece of memory being freed in a destructor. Bug was reported by
Santiago Escobar <sescobar@dsic.upv.es> but the example is too big to
include in the release note.

There is no PowerPC/Darwin binary at the moment because our
PowerPC/Darwin server has a hard drive failure.

Steven