Maude Alpha89e
Dear Maude Abusers,

These is bug fix release to fix a critical problem in the garbage
collection of unification subproblems exposed by an example from
Santiago. A simplified example to provoke the problem is:

fmod FOO is
  sort Foo .
  ops a b : -> Foo .
  op f : Foo Foo -> Foo [assoc comm] .
  op g : Foo Foo -> Foo .
  vars W X Y Z : Foo .
endfm

unify g(f(X, Y), a) =? g(f(W, Z), b) .

What is happening is that f(X, Y) =? f(W, Z) created a unification
problem then a =? b causes immediate failure and deletion of all
subproblems generated so far; except that a pointer to the last such
subproblem doesn't get zeroed as that pointer was being used in two
different ways and the last subproblem gets deleted a second time when
the rest of the unification problem is deleted.

Steven