Maude Alpha91d
Dear Maude Abusers,

This release exists to fix a critical bug reported by Marcel Kyas
<kyas@ifi.uio.no>

The problem is with the normalization of terms in the AU theory,
either during compilation of a module or before a term is converted to
a DAG for rewriting. In the case that all of the arguments to an AU
function symbol are identity elements, the term is collapsed to the
first such element after it has been deleted, leading to memory
corruption and a likely segmentation fault. Here is an example that
triggers the problem.

fmod FOO is
  sort Foo .
  op 1 : -> Foo .
  op f : Foo Foo -> Foo [assoc id: 1] .
endfm

red f(1, 1) .

This bug has existed since the AU theory was supported.

Steven